Conflicts:
	source/algorithms/randompermutation/RandomPermutation.h
	source/structures/projection/ProjectionLimits.hpp
	src/libalgo/source/algorithms/cartanalysis/CartAnalysis_batch.hpp.bak
	src/libalgo/source/algorithms/linelineposition/LineLinePosition.h.bak
	src/libalgo/source/algorithms/linelineposition/LineLinePosition.hpp.bak
	src/libalgo/source/algorithms/matrixoperations/MatrixOperations.h_bak2
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjB3R.cpp
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjB3R.h
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjJ2.h.bak2
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjJ3.h.bak2
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjJ3R.cpp
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjJ3R.h
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjJ4.cpp.bak
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjJ4.h.bak
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjV2.h.bak2
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjV3.h.bak2
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjV3R.cpp
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjV3R.h
	src/libalgo/source/algorithms/nonlinearleastsquares/FAnalyzeProjV4.h.bak
	src/libalgo/source/algorithms/numderivative/FProjEquationDerivative6VarR.cpp
	src/libalgo/source/algorithms/numderivative/FProjEquationDerivative6VarR.h
	src/libalgo/source/algorithms/turningfunction/TurningFunction.h.bak2
	src/libalgo/source/algorithms/turningfunction/TurningFunction.hpp.bak
	src/libalgo/source/algorithms/turningfunction/TurningFunction.hpp.bak2
	src/libalgo/source/algorithms/turningfunction/TurningFunction.hpp.bak3
	src/libalgo/source/comparators/sortSamplesByAllRatios.h.bak
	src/libalgo/source/structures/matrix/Matrix.h.bak
	src/libalgo/source/structures/matrix/Matrix.hpp.bak
	src/libalgo/source/structures/projection/Limits.cpp
	src/libalgo/source/structures/projection/Limits.h
	src/libalgo/source/structures/projection/ProjectionAzimuthalLimits.cpp
	src/libalgo/source/structures/projection/ProjectionAzimuthalLimits.h
	src/libalgo/source/structures/projection/ProjectionConicLimits.cpp
	src/libalgo/source/structures/projection/ProjectionConicLimits.h
	src/libalgo/source/structures/projection/ProjectionCylindricalLimits.cpp
	src/libalgo/source/structures/projection/ProjectionCylindricalLimits.h
	src/libalgo/source/structures/projection/ProjectionEllipsoidalLimits.cpp
	src/libalgo/source/structures/projection/ProjectionEllipsoidalLimits.h
	src/libalgo/source/structures/projection/ProjectionLimits.cpp
	src/libalgo/source/structures/projection/ProjectionLimits.h
	src/libalgo/source/structures/projection/ProjectionLimits.hpp
	src/libalgo/source/structures/projection/ProjectionMiscellaneousLimits.cpp
	src/libalgo/source/structures/projection/ProjectionMiscellaneousLimits.h
	src/libalgo/source/structures/projection/ProjectionPolyConicLimits.cpp
	src/libalgo/source/structures/projection/ProjectionPolyConicLimits.h
	src/libalgo/source/structures/projection/ProjectionPseudoAzimuthalLimits.cpp
	src/libalgo/source/structures/projection/ProjectionPseudoAzimuthalLimits.h
	src/libalgo/source/structures/projection/ProjectionPseudoConicLimits.cpp
	src/libalgo/source/structures/projection/ProjectionPseudoConicLimits.h
	src/libalgo/source/structures/projection/ProjectionPseudoCylindricalLimits.cpp
	src/libalgo/source/structures/projection/ProjectionPseudoCylindricalLimits.h

+++ src/libalgo/source/algorithms/simplexmethod/SimplexMethod.h
@@ -31,7 +31,7 @@ class SimplexMethod
         public:
 
                 template <typename T, typename Function>
-                static T NelderMead ( Function function, const Matrix <T> &XMIN, const Matrix <T> &XMAX, Matrix <T> &W, Matrix <T> &X, Matrix <T> &Y, Matrix <T> &V, unsigned short &iterations, const T max_error, const unsigned short max_iterations, std::ostream * output = &std::cout );
+                static T NelderMead ( Function function, const Matrix <T> &XMIN, const Matrix <T> &XMAX, Matrix <T> &W, Matrix <T> &X, Matrix <T> &Y, Matrix <T> &V, unsigned int &iterations, const T max_error, const unsigned int max_iterations, std::ostream * output = &std::cout );
 
         private:
